+++ CHANGELOG.md
@@ -8,3 +8,4 @@
 
 - [#44](https://github.com/wallix/awless/issues/44): Fetch only the S3 buckets and related objects of the current region.
 - [#52](https://github.com/wallix/awless/issues/52)-[#34](https://github.com/wallix/awless/issues/34): Properly fetch route tables, even if a route contains several destinations.
+- [#37](https://github.com/wallix/awless/issues/37): Load the region from database when initializing cloud services rather than `awless` environment.